Rename ScriptsPage to ScriptIndexPage
