+++ frontend/src/AuthenticatedApp/paperScripts/ScriptIndexPage/ScriptsTableRow.tsx
