+++ frontend/src/AuthenticatedApp/paperScripts/ScriptIndexPage/index.tsx
