+++ frontend/src/AuthenticatedApp/paperScripts/index.ts
@@ -1,6 +1,6 @@
+import ScriptIndexPage from "./ScriptIndexPage";
 import ScriptMarkSubpage from "./ScriptMarkSubpage";
-import ScriptsPage from "./ScriptsPage";
 import ScriptViewSubpage from "./ScriptViewSubpage";
 
 export { ScriptMarkSubpage, ScriptViewSubpage };
-export default ScriptsPage;
+export default ScriptIndexPage;
